Address comments for pr(10)
